ai/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# OpenAI and AI
openai==1.54.0
chromadb==0.5.20

# Data handling
pydantic==2.10.0
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Development
ruff==0.7.4
